doctor-checkup

+++ src/main/java/org/pahappa/systems/models/User.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    int id;
+    Long id;
     private String fullName;
     private String phoneNumber;
     private String email;
@@ -44,11 +44,11 @@ public class User {
 
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
